+++ R/named.R
@@ -1,6 +1,7 @@
 #' Check Named
 #'
 #' @param x The object to check.
+#' @param regex A string of the regular expression that must match all names.
 #' @param unique A flag indicating whether the names must be unique.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
@@ -12,7 +13,7 @@
 #' x <- 1
 #' names(x) <- "y"
 #' check_named(x, error = FALSE)
-check_named <- function(x, unique = FALSE,
+check_named <- function(x, regex = ".*", unique = FALSE, 
                          x_name = substitute(x),
                          error = TRUE) {
   x_name <- deparse_x_name(x_name)
@@ -22,7 +23,7 @@ check_named <- function(x, unique = FALSE,
 
   if(is.null(names(x))) {
     on_fail(x_name, " must be named", error = error)
-  }
+  } else check_regex(names(x), regex = regex, x_name = paste("names of", x_name), error = error)
   if(unique)
     check_unique(names(x), x_name = paste("names of", x_name), error = error)
   invisible(x)